requirements.txt
# ML
tensorflow>=2.19.0
keras>=3.0.0
numpy>=1.26.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Load Testing
locust>=2.19.0

# Data
kagglehub
